refactor(test): move SI test data creation to erpnext test suite
